+++ test/ModelMixins/CatalogMemberMixinSpec.ts
@@ -2,6 +2,11 @@ import { runInAction, IReactionDisposer, reaction } from "mobx";
 import Terria from "../../lib/Models/Terria";
 import WebMapServiceCatalogItem from "../../lib/Models/WebMapServiceCatalogItem";
 import updateModelFromJson from "../../lib/Models/updateModelFromJson";
+import CommonStrata from "../../lib/Models/CommonStrata";
+import createStratumInstance from "../../lib/Models/createStratumInstance";
+import DimensionTraits, {
+  DimensionOptionTraits
+} from "../../lib/Traits/DimensionTraits";
 
 describe("CatalogMemberMixin", function() {
   describe(" - infoWithoutSources", function() {
@@ -121,6 +126,52 @@ describe("CatalogMemberMixin", function() {
       expect(wmsItem.isLoading).toBeFalsy();
       expect(wmsItem.isLoadingMetadata).toBeFalsy();
       expect(wmsItem.isLoadingMapItems).toBeFalsy();
+    });
+
+    it(" - modelDimensions", () => {
+      wmsItem.setTrait(CommonStrata.definition, "styles", "init-style");
+      wmsItem.setTrait(CommonStrata.definition, "layers", "init-layers");
+      wmsItem.setTrait(CommonStrata.user, "modelDimensions", [
+        createStratumInstance(DimensionTraits, {
+          id: "modelDimensions",
+          options: [
+            createStratumInstance(DimensionOptionTraits, {
+              id: "styles-test",
+              value: { styles: "test" }
+            }),
+            createStratumInstance(DimensionOptionTraits, {
+              id: "styles-test2",
+              value: { styles: "test2" }
+            }),
+            createStratumInstance(DimensionOptionTraits, {
+              id: "layers-test",
+              value: { layers: "test" }
+            })
+          ]
+        })
+      ]);
+
+      expect(wmsItem.styles).toBe("init-style");
+      expect(wmsItem.layers).toBe("init-layers");
+
+      const modelDimension = wmsItem.selectableDimensions.find(
+        dim => dim.id === "modelDimensions"
+      );
+
+      modelDimension?.setDimensionValue(CommonStrata.user, "styles-test");
+
+      expect(wmsItem.styles).toBe("test");
+      expect(wmsItem.layers).toBe("init-layers");
+
+      modelDimension?.setDimensionValue(CommonStrata.user, "styles-test2");
+
+      expect(wmsItem.styles).toBe("test2");
+      expect(wmsItem.layers).toBe("init-layers");
+
+      modelDimension?.setDimensionValue(CommonStrata.user, "layers-test");
+
+      expect(wmsItem.styles).toBe("test2");
+      expect(wmsItem.layers).toBe("test");
     });
   });
 });